packages: util-linux: add host build for libuuid
This patch adds a host build for util-linux to make libuuid available in
the host environment.

libuuid is needed as a build dependency of gupnp, which needs a host
build for the latest version in order to do Vala API generation.

Signed-off-by: Andy Leiserson <andy@leiserson.org>
---
To me, this approach seemed preferable to making libuuid a hard prereq
for the openwrt build (i.e. include/prereq-build.mk). However, I'm happy
to submit a patch for the latter if desired.